Refactor lambda functions
