+++ src/functions/slack-api/create-conversation.js
@@ -1,27 +1,32 @@
 const {botApp, awsLambdaReceiver} = require("../../helpers/aws-slack-bot");
 
-// Handle the Lambda function event
 module.exports.handler = async (event, context, callback) => {
     const handler = await awsLambdaReceiver.start();
 
     const data = JSON.parse(event.body);
 
     try {
-        await botApp.client.chat.postMessage({
-            channel: data.channel,
-            text: data.text
+        const result = await botApp.client.conversations.open({
+            users: data.users.join()
         });
-
+        const responseBody = toResponseBody(result);
         callback(null, {
             statusCode: 200,
-            body: "Message send."
+            body: JSON.stringify(responseBody)
         })
     } catch (error) {
         callback(null, {
             statusCode: error.statusCode || 500,
-            body: "Could not send message. Reason is:\n " + error
+            body: `Could not create conversation. Reason is:\n\t${error}`
         })
     }
 
     return handler(event, context, callback);
+}
+
+function toResponseBody(result) {
+    return {
+        id: result.channel.id,
+        name: result.channel.name,
+    };
 }
